main: use one entry point for all protocols, implement stack overflow test
Some operations are required for all protocols, like TPM and event log
setup or memory protection initialization (to be done). It is easier and
safer to change one common function than to repeat work in every protocol.

This change, on top of TPM event log implementation, uses more stack that
was previously allocated. SL header and probably some memory just below it
was overwritten by local variables and other temporary data. LZ is still
able to finish and pass control to the kernel, but the kernel cannot find
the LZ header, because its offset was overwritten. This also potentially
can overwrite code of other components, if they are loaded directly below
LZ.

Apart from incrementing stack size, add a stack canary and reboot if it
was changed. Hopefully, this will help to avoid wasting time debugging
wrong component.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>